// frontend/src/backend.d.ts
export interface Transaction {
    id: bigint;
    paymentMethod?: PaymentMethod;
    transactionType: TransactionType;
    date: bigint;
    description: string;
    amount: bigint;
    expenseCategory?: ExpenseCategory;
}
// ...
export enum ExpenseCategory {
    rezije = "rezije",
    oprema = "oprema",
    napojnica = "napojnica",
    pice = "pice",
    place = "place",
    ostalo = "ostalo",
    namirnice = "namirnice"
}
export enum OfferType {
    bar = "bar",
    brzaHrana = "brzaHrana",
    kafic = "kafic",
    ostalo = "ostalo",
    restoran = "restoran"
}
export enum PaymentMethod {
    kartica = "kartica",
    gotovina = "gotovina"
}
export enum SeasonalActivity {
    oboje = "oboje",
    zima = "zima",
    ljeto = "ljeto"
}
export enum TransactionType {
    prihod = "prihod",
    rashod = "rashod"
}